Only admin role can delete article.

+++ templates/articles/view.php
@@ -2,5 +2,8 @@
     <h1><?= $article->getName() ?></h1>
     <p><?= $article->getParsedText() ?></p>
     <p>Автор: <?= $article->getAuthor()->getNickname() ?></p>
-    <a href="/articles/<?=$article->getId() ?>/delete">Delete</a>
+
+    <?php if (!empty($user) && $user->isAdmin() ) {?>
+        <a href="/articles/<?=$article->getId() ?>/delete">Delete</a>
+    <?php }?>
 <?php include __DIR__ . '/../footer.php'; ?>
